+++ src/main/java/racingcar/Application.java
@@ -1,5 +1,7 @@
 package racingcar;
 
+import racingcar.controller.RacingGame;
+
 public class Application {
     public static void main(String[] args) {
         RacingGame racingGame = new RacingGame();
