+++ src/main/java/racingcar/model/Car.java
@@ -1,4 +1,6 @@
-package racingcar;
+package racingcar.model;
+
+import racingcar.util.RandomNumberGenerator;
 
 public class Car {
     private String name;
